% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/protractedbirthdeath.R
\name{combinelists}
\alias{combinelists}
\title{An internal function to quickly combine lists}
\usage{
combinelists(x)
}
\arguments{
\item{x}{a list}
}
\value{
a dataframe containing sensibly formatted data for use in \code{\link{repsim2}}
}
\description{
An internal function to quickly combine lists
}

