% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/protractedbirthdeath.R
\name{countpersistance}
\alias{countpersistance}
\title{Finds persistance of taxa over time.
gives a number of taxa alive at t1 and t2 or both}
\usage{
countpersistance(df, t1 = 10, t2 = 5)
}
\arguments{
\item{df}{single run from repsim2}

\item{t1}{first time point}

\item{t2}{second time point}
}
\value{
a vector of 3 values - those alive at t1, both and t2
}
\description{
Finds persistance of taxa over time.
gives a number of taxa alive at t1 and t2 or both
}
\examples{
\dontrun{
countpersistance(repsim2(c(0.2,0.2,0.2,0.1,0.1),1,15),10,5)
}
}
\seealso{
\code{\link{repsim2}} which produces the inputs
}

