% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/protractedbirthdeath.R
\name{dplyframe}
\alias{dplyframe}
\title{dplyframe
an internal function to determine living species at each time point}
\usage{
dplyframe(x)
}
\arguments{
\item{x}{a single run of loopfunct}
}
\value{
a dataframe with mean and sd number of species in each class at each integer time point run over all repeats
}
\description{
dplyframe
an internal function to determine living species at each time point
}
\seealso{
\code{\link{summaryrepsim}} which loops this over multiple runs for means and sds
}

