% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/protractedbirthdeath.R
\name{loopfunct}
\alias{loopfunct}
\title{loopfunct
an internal function to determine living species at each time point}
\usage{
loopfunct(x, time = 15)
}
\arguments{
\item{x}{a single run of repsim2}

\item{time}{the time repsim2 was run for}
}
\value{
a dataframe with number of species in each class at each integer time point run over all repeats
}
\description{
loopfunct
an internal function to determine living species at each time point
}
\seealso{
\code{\link{summaryrepsim}} which loops this over multiple runs for means and sds
}

