% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/protractedbirthdeath.R
\name{numgoodincip}
\alias{numgoodincip}
\title{Finds numbers of each possible outcome at end of simulation.}
\usage{
numgoodincip(df)
}
\arguments{
\item{a}{single run from repsim2}
}
\value{
a vector of 4 named numbers - dead and live good and incipient taxa
}
\description{
Finds numbers of each possible outcome at end of simulation.
}
\examples{
numgoodincip(repsim2(c(0.2,0.2,0.2,0.1,0.1),1,15))
}
\seealso{
\code{\link{repsim2}} which produces the inputs
}

