% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protractedbirthdeath.R
\name{pbdshiny}
\alias{pbdshiny}
\title{A shiny app to view the simulations}
\usage{
pbdshiny()
}
\value{
a shiny app
}
\description{
A shiny app to view the simulations
}
\examples{
\dontrun{
pbdshiny()
}
}

