% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/protractedbirthdeath.R
\name{pbdsim2}
\alias{pbdsim2}
\title{Simulate speciation by a protracted birth death process.}
\usage{
pbdsim2(pars, totaltime = 15)
}
\arguments{
\item{a}{vector of 5 parameters, and a time: c(speciation rate of good species,
speciation rate of incipient species, completion rate, death rate of
good species and death rate of incipient species). The time is following}
}
\value{
a list with each taxa as its own entry, with birth and death times
  has an atrribute overloaded=1 if too many species were formed
}
\description{
Simulate speciation by a protracted birth death process.
}
\examples{
pbdsim2(c(0.1,0.1,0.1,0.1,0.1),15)
pbdsim2(c(0.2,0.2,0.2,0.1,0.1),15)
}
\seealso{
\code{\link{repsim2}} which wraps this function and does repeats
}

