% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protractedbirthdeath.R
\name{plotcontour}
\alias{plotcontour}
\title{A contourplot of number of species across differing parameters
chooses a variable - mean sd of alltaxa, good taxa etc
and plots it for fixed values of three chosen variables}
\usage{
plotcontour(data, variable, xx, yy, var1, val1, var2, val2, var3, val3,
  logged = F, numbins)
}
\arguments{
\item{data}{a loaded frame of multiple results see example data for formatting}

\item{variable}{a value to plot - usually mean}

\item{xx}{the x axis variable}

\item{yy}{the y axis variable}

\item{var1}{the name of variable 1}

\item{val1}{the selected value of variable 1}

\item{var2}{the name of variable 2}

\item{val2}{the selected value of variable 2}

\item{var3}{the name of variable 3}

\item{val3}{the selected value of variable 3}

\item{logged}{whether to plot logged data, defaults to F}

\item{numbins}{the number of bins to use in plotting contours}
}
\value{
a ggplot with the two non named variables and countours for the chosen variable
}
\description{
A contourplot of number of species across differing parameters
chooses a variable - mean sd of alltaxa, good taxa etc
and plots it for fixed values of three chosen variables
}
\seealso{
\code{\link{repsim2}} which produces the inputs
}

