% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protractedbirthdeath.R
\name{removeyoungest}
\alias{removeyoungest}
\title{Remove youngest taxa from tree
the recurssive function used to prune trees for tree construction in
\code{\link{treemaker2}}. Works by taking the youngest offspring from each parent,
making sure it isnt a parent. If it isnt, it is trimmed and removed from the list
and its parent is relabelled. Probably my hardest thought out function.
Significantly faster than simply removing the youngest taxa as this prunes all at once.}
\usage{
removeyoungest(z)
}
\arguments{
\item{z}{a single run from repsim2}
}
\value{
dataframe with the youngest offsprings all removed
}
\description{
Remove youngest taxa from tree
the recurssive function used to prune trees for tree construction in
\code{\link{treemaker2}}. Works by taking the youngest offspring from each parent,
making sure it isnt a parent. If it isnt, it is trimmed and removed from the list
and its parent is relabelled. Probably my hardest thought out function.
Significantly faster than simply removing the youngest taxa as this prunes all at once.
}
\examples{
\dontrun{
x<-removeyoungest(repsim2(c(0.2,0.2,0.2,0.1,0.1),1,15))
}
}
\seealso{
\code{\link{repsim2}} which produces the inputs
}

