% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protractedbirthdeath.R
\name{sisterlengths}
\alias{sisterlengths}
\title{Finds branchlength to closest extant sister taxa.}
\usage{
sisterlengths(working)
}
\arguments{
\item{a}{single run from repsim2}
}
\value{
a list of branch lengths
}
\description{
Finds branchlength to closest extant sister taxa.
}
\examples{
\dontrun{
sisterlengths(repsim2(c(0.2,0.2,0.2,0.1,0.1),1,15))
}
}
\seealso{
\code{\link{repsim2}} which produces the inputs
}

