% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/protractedbirthdeath.R
\name{subsetdata}
\alias{subsetdata}
\title{subsetdata
subsets data based on parameters}
\usage{
subsetdata(data, var1, val1, var2, val2, var3, val3)
}
\arguments{
\item{data}{a loaded frame of multiple results see example data for formatting}

\item{var1}{the name of variable 1}

\item{val1}{the selected value of variable 1}

\item{var2}{the name of variable 2}

\item{val2}{the selected value of variable 2}

\item{var3}{the name of variable 3}

\item{val3}{the selected value of variable 3}
}
\description{
subsetdata
subsets data based on parameters
}

