% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/protractedbirthdeath.R
\name{sumfunct}
\alias{sumfunct}
\title{sumfunct
an internal function to determine living species at each time point}
\usage{
sumfunct(df, time = 15)
}
\arguments{
\item{df}{a single run of repsim2}

\item{time}{the time repsim2 was run for}
}
\value{
a dataframe with number of species in each class at each integer time point
}
\description{
sumfunct
an internal function to determine living species at each time point
}
\seealso{
\code{\link{summaryrepsim}} which loops this over multiple runs for means and sds
}

