% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/protractedbirthdeath.R
\name{sumfunctpart1}
\alias{sumfunctpart1}
\title{sumfunctpart1
an internal function to determine living species at each time point
run across (apply,1) on a repsim result to give each lines time of death etc}
\usage{
sumfunctpart1(x, time = 15)
}
\arguments{
\item{x}{a single line of a run from repsim2}

\item{time}{the time the simulation was run for}
}
\value{
a vector with 0,1,2,3,4 at each integer t depending on taxa state
summed to give summaries
}
\description{
sumfunctpart1
an internal function to determine living species at each time point
run across (apply,1) on a repsim result to give each lines time of death etc
}
\seealso{
\code{\link{summaryrepsim}} which links everything together
}

