% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/protractedbirthdeath.R
\name{sumfunctpart2}
\alias{sumfunctpart2}
\title{sumfunctpart2
an internal function to determine living species at each time point
run across (apply,1) on a sumfunctpart1 result to give each lines time of death etc}
\usage{
sumfunctpart2(downcol)
}
\arguments{
\item{downcol}{a single line of a run from sumfunct1}
}
\value{
a matrix with taxa state for each n
}
\description{
sumfunctpart2
an internal function to determine living species at each time point
run across (apply,1) on a sumfunctpart1 result to give each lines time of death etc
}
\seealso{
\code{\link{summaryrepsim}} which links everything together
}

