% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/protractedbirthdeath.R
\name{summaryrepsim}
\alias{summaryrepsim}
\title{A summary of repeats with means and sds.}
\usage{
summaryrepsim(pars, n, time)
}
\arguments{
\item{a}{vector of 5 parameters, a repeat number and a time:
c(speciation rate of good species,
speciation rate of incipient species, completion rate, death rate of
good species and death rate of incipient species). The time is following}
}
\value{
a data frame with means and sds of numbers of taxa at each timepoint
}
\description{
A summary of repeats with means and sds.
}
\examples{
summaryrepsim(c(0.2,0.2,0.2,0.1,0.1),15,15)
}
\seealso{
\code{\link{repsim2}} which produces the inputs \code{\link{plotsim}}
  which plots this functions output
}

