% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/protractedbirthdeath.R
\name{tau}
\alias{tau}
\title{calculates tau - time to complete speciation}
\usage{
tau(l2, l3, m2)
}
\arguments{
\item{l2}{speciation completion rate}

\item{l3}{incipient speciation rate}

\item{m2}{incipient species death rate}
}
\value{
a single tau value
}
\description{
calculates tau - time to complete speciation
}

