% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/protractedbirthdeath.R
\name{taufunct}
\alias{taufunct}
\title{calculates tau with fixed values
allows choosing of variables to plot}
\usage{
taufunct(var2, x, y, fixed)
}
\arguments{
\item{var2}{variable to fix}

\item{x}{first other variable}

\item{y}{second other variable}

\item{fixed}{the fixed value to give var2}
}
\value{
a single tau value
}
\description{
calculates tau with fixed values
allows choosing of variables to plot
}

