% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protractedbirthdeath.R
\name{treemaker}
\alias{treemaker}
\title{Takes a single inputted run and makes a (text) tree.
deprecated - much slower than treemaker2}
\usage{
treemaker(x)
}
\arguments{
\item{x}{a single run from repsim2}
}
\value{
a string which can be read into ape using read.tree
}
\description{
Takes a single inputted run and makes a (text) tree.
deprecated - much slower than treemaker2
}
\examples{
\dontrun{
x<-treemaker(repsim2(c(0.2,0.2,0.2,0.1,0.1),1,15))
phy<-read.tree(text=x)
plot(phy)
}
}
\seealso{
\code{\link{repsim2}} which produces the inputs

\code{\link{treemaker2}} a much faster implementation
}

