% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/protractedbirthdeath.R
\name{countpersistance}
\alias{countpersistance}
\title{Finds persistance of taxa over time.}
\usage{
countpersistance(df, t1 = 10, t2 = 5)
}
\arguments{
\item{a}{single run from repsim2, and two timepoints to check}
}
\value{
a vector of 3 values - those alive at t1, both and t2
}
\description{
Finds persistance of taxa over time.
}
\examples{
countpersistance(repsim2(c(0.2,0.2,0.2,0.1,0.1),1,15),10,5)
}
\seealso{
\code{\link{repsim2}} which produces the inputs
}

