% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/protractedbirthdeath.R
\name{plotcontour}
\alias{plotcontour}
\title{A contourplot of number of species across differing parameters}
\usage{
plotcontour(data, variable, xx, yy, var1, val1, var2, val2, var3, val3,
  logged = F, numbins)
}
\arguments{
\item{dataframe}{with many simulations, the variable to plot (ie good species)
then the 3 fixed parameters and the values wanted for the plot}
}
\value{
a ggplot with the two non named variables and countours for the chosen variable
}
\description{
A contourplot of number of species across differing parameters
}
\seealso{
\code{\link{repsim2}} which produces the inputs
}

