% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protractedbirthdeath.R
\name{summaryrepsim}
\alias{summaryrepsim}
\title{summaryrepsim
A summary of repeats with means and sds.}
\usage{
summaryrepsim(pars, n, time)
}
\arguments{
\item{pars}{a vector of 5 parameters: c(speciation rate of good species,
speciation rate of incipient species, completion rate, death rate of
good species and death rate of incipient species).}

\item{n}{a number of times to repeat the simulation}

\item{time}{time to run simulation, defaults to 15}
}
\value{
a data frame with means and sds of numbers of taxa at each timepoint
}
\description{
summaryrepsim
A summary of repeats with means and sds.
}
\examples{
\dontrun{
summaryrepsim(c(0.2,0.2,0.2,0.1,0.1),15,15)
}
}
\seealso{
\code{\link{repsim2}} which produces the inputs \code{\link{plotsim}}
  which plots this functions output
}

