% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protractedbirthdeath.R
\name{tauloop}
\alias{tauloop}
\title{A matrix of tau values.}
\usage{
tauloop(x, var)
}
\arguments{
\item{a}{value to fix, and the variable it is}
}
\value{
a matrix of tau, read to plot
}
\description{
A matrix of tau values.
}
\examples{
tauloop(0.5,"incipext")
}
\seealso{
\code{\link{plottau}} which plots the outputs
}

