% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/protractedbirthdeath.R
\name{timegivengood}
\alias{timegivengood}
\title{Finds time to be a good species.}
\usage{
timegivengood(df)
}
\arguments{
\item{a}{single run from repsim2}
}
\value{
a vector of times - equal to number of good species
}
\description{
Finds time to be a good species.
}
\examples{
timegivengood(repsim2(c(0.2,0.2,0.2,0.1,0.1),1,15),10,5)
}
\seealso{
\code{\link{repsim2}} which produces the inputs
}

