% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/protractedbirthdeath.R
\name{treemaker}
\alias{treemaker}
\title{Takes a single inputted run and makes a (text) tree.}
\usage{
treemaker(x)
}
\arguments{
\item{a}{single run from repsim2}
}
\value{
a string which can be read into ape using read.tree
}
\description{
Takes a single inputted run and makes a (text) tree.
}
\examples{
x<-treemaker(repsim2(c(0.2,0.2,0.2,0.1,0.1),1,15))
phy<-read.tree(text=x)
plot(phy)
}
\seealso{
\code{\link{repsim2}} which produces the inputs
}

